ai: port simple-options and transform-messages (task 8, part 2)
Both modules are shared by every API implementation, so they land before the
Anthropic protocol itself.

- simple_options: context clamping (4 096 safety tokens, at least one token
  left), buildBaseOptions with per-key sampling-param merge, and the thinking
  budgets (minimal 1024 / low 2048 / medium 8192 / high 16384) including the
  rule that a budget never eats the last MIN_ANSWER_TOKENS.
- transform_messages: image downgrade for text-only models with collapsed
  placeholder runs, thinking-block handling across models (redacted payloads
  dropped, signed blocks kept, others converted to text), tool-call id
  normalization that rewrites the matching tool results, dropping of errored
  and aborted turns, and synthetic results for orphaned tool calls.

Token widths are unified to u64 because JS has a single number type and
maxTokens is compared against model.maxTokens and contextWindow; noted for the
other workstreams as interface-request B-3.

Tests (14) include the exact 4 899-token clamp of the TS context-estimate
fixture, which caught that my first fixture had left out the stale assistant
message.

+++ crates/notagent-ai/src/types.rs
@@ -197,13 +197,13 @@ pub enum ChatTemplateKwargValue {
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Default, Serialize, Deserialize)]
 pub struct ThinkingBudgets {
     #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub minimal: Option<u32>,
+    pub minimal: Option<u64>,
     #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub low: Option<u32>,
+    pub low: Option<u64>,
     #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub medium: Option<u32>,
+    pub medium: Option<u64>,
     #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub high: Option<u32>,
+    pub high: Option<u64>,
 }
 
 // ---------------------------------------------------------------------------
@@ -349,7 +349,8 @@ pub struct StreamOptions {
     /// Wird nach den benannten Feldern in den Request-Body gemischt und überschreibt sie
     /// dadurch; nur OpenAI-kompatible Adapter werten sie aus.
     pub sampling_params: Option<Map<String, Value>>,
-    pub max_tokens: Option<u32>,
+    /// Token counts share one width across model, context and request options.
+    pub max_tokens: Option<u64>,
     pub transport: Option<Transport>,
     /// Default: `"short"`.
     pub cache_retention: Option<CacheRetention>,
